refactor(middleware): split into multiple files (#664)
